+++ catalog/controller/extension/payment/duitku_va_danamonh2h.php
@@ -147,9 +147,9 @@ class ControllerExtensionPaymentDuitkuVADanamonh2h extends Controller {
       $this->cart->clear();
 	  $this->model_checkout_order->addOrderHistory($order_id, $this->config->get('payment_duitku_va_danamonh2h_pending_mapping'), 'Duitku payment pending.');
       				
-      $this->log->write("Request : " . json_encode($params) );		
-      $redirUrl = DuitkuCore_Web::getRedirectionUrl($this->config->get('payment_duitku_va_danamonh2h_endpoint'), $params);
-      $this->response->setOutput($redirUrl);	  
+      $this->log->write("Request : " . json_encode($params, JSON_PRETTY_PRINT) );		
+      $redirUrl = DuitkuCore_Web::getRedirectionUrl($this->config->get('payment_duitku_va_danamonh2h_endpoint'), $params,  $this->log);
+      $this->response->setOutput($redirUrl);
     }
     catch (Exception $e) {
       $data['errors'][] = $e->getMessage();
@@ -162,12 +162,12 @@ class ControllerExtensionPaymentDuitkuVADanamonh2h extends Controller {
    * Landing page when payment is finished or failure or customer pressed "back" button
    * The Cart is cleared here, so make sure customer reach this page to ensure the cart is emptied when payment succeed
    * payment finish/unfinish/error url :
-   * http://[your shop’s homepage]/index.php?route=payment/duitku_va_danamonh2h/payment_notification
+   * http://[your shopï¿½s homepage]/index.php?route=payment/duitku_va_danamonh2h/payment_notification
    */
   public function landing_redir() {    
     $this->load->model('checkout/order');
     $this->load->model('extension/payment/duitku_va_danamonh2h');    
-    $redirUrl = $this->url->link('checkout/cart');
+    $redirUrl = $this->url->link('checkout/success');
 
     if (isset($_GET['resultCode']) && isset($_GET['merchantOrderId']) && isset($_GET['reference']) && $_GET['resultCode'] == '01') {
       //if capture or pending or challenge or settlement, redirect to order received page
@@ -252,11 +252,28 @@ class ControllerExtensionPaymentDuitkuVADanamonh2h extends Controller {
     $merchant_code = $this->config->get('payment_duitku_va_danamonh2h_merchant');    
     $endpoint = $this->config->get('payment_duitku_va_danamonh2h_endpoint');
 
+    $signatureCheck = md5($merchant_code . intval($_REQUEST['amount']) . $_REQUEST['merchantOrderId'] . $api_key);
+
+    $order_info = $this->model_checkout_order->getOrder($order_id);
+    $current_status_id = $order_info['order_status_id'];
+
+    if ($current_status_id == $this->config->get('payment_duitku_va_danamonh2h_success_mapping')){
+      header("HTTP/1.1 200");
+      echo "Order Already Completed";
+      die;
+    }
+
+    if ($_REQUEST['signature'] != $signatureCheck){
+      header("HTTP/1.1 500 Internal Server Error");
+      echo "Wrong Signature";
+      die;
+    }
+
     $order_info = $this->model_checkout_order->getOrder($order_id);
 
     //check if order id is in the database
     if ($order_info) {
-        $this->log->write("perform validation");
+      $this->log->write("Callback Recieved : " . json_encode($_REQUEST, JSON_PRETTY_PRINT));
         if ($status == '00' && DuitkuCore_Web::validateTransaction($endpoint, $merchant_code, $order_id, $reference, $api_key)) {
             $this->model_checkout_order->addOrderHistory($order_id, $this->config->get('payment_duitku_va_danamonh2h_success_mapping'), 'Duitku payment successful.');    
         } else {